refactor(helpers): scope for single use helpers
