+++ src/components/Pagination/helpers.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { range } from './range';
+import { range } from './helpers';
 
 describe('Helpers / Range', () => {
   it('should return the empty list, given start >= end', () => {
